feat(inventory): support fractional quantities in stock deduction ACL
Widen StockDeductionItem.quantity from int to double so the InventoryContextFacade
can deduct weight-based products in kilograms with decimals. Unit products keep whole
quantities via rounding at the unit-lot deduction call. Existing callers that pass
integer quantities remain source-compatible through implicit widening.

+++ src/main/java/online/entreprenly/platform/inventory/interfaces/acl/StockDeductionItem.java
@@ -5,7 +5,8 @@ package online.entreprenly.platform.inventory.interfaces.acl;
  * contexts through the {@link InventoryContextFacade}.
  *
  * @param productName the product display name, matched case-insensitively
- * @param quantity    the quantity to deduct (units for unit products, kilograms for weight products)
+ * @param quantity    the quantity to deduct (whole units for unit products, kilograms —
+ *                    fractional allowed — for weight products)
  */
-public record StockDeductionItem(String productName, int quantity) {
+public record StockDeductionItem(String productName, double quantity) {
 }
